backend/pyproject.toml
[project]
name = "vandalizer-next-backend"
version = "0.1.0"
description = "Vandalizer Next - FastAPI backend"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "beanie>=1.27,<2",
    "motor>=3.6,<4",
    "pydantic-settings>=2.7,<3",
    "python-jose[cryptography]>=3.3,<4",
    "werkzeug>=3.1,<4",
    "python-multipart>=0.0.18,<1",
    "celery[redis]>=5.4,<6",
    "pydantic-ai>=0.1",
    "pymongo>=4.0,<5",
    "chromadb>=0.5,<0.6",
    "langchain-text-splitters>=0.3,<1",
    "langchain-openai>=0.3,<1",
    "httpx>=0.27,<1",
    "beautifulsoup4>=4.12,<5",
    "aiosmtplib>=3.0,<4",
    "PyPDF2>=3.0,<4",
    "pypandoc>=1.13,<2",
    "markitdown>=0.1,<1",
    "markdownify>=0.13,<1",
    "msal>=1.28,<2",
    "redis>=5.0,<6",
    "slowapi>=0.1,<1",
    "fpdf2>=2.8.7,<3",
    "sentry-sdk[fastapi,celery]>=2.0,<3",
    "python-json-logger>=3.0,<4",
    "python3-saml>=1.16,<2",
    "asteval>=0.9,<1",
    "prometheus-fastapi-instrumentator>=7.0,<8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
